refactor: migrate PostgreSQL resource change extraction to ANTLR parser
This commit completes the migration from pg_query_go to ANTLR-based parser
for extracting changed resources in PostgreSQL SQL statements.

Key changes:
- Consolidated resource change logic into resource_change.go using ANTLR
- Removed legacy resource_change_antlr.go file
- Implemented changedResourcesListener to handle all DDL and DML statements
- Fixed range calculation to include trailing semicolons in SQL statements

The ANTLR implementation provides better maintainability and consistency
with other parser functionality while maintaining full compatibility with
existing behavior.
